A bulk of things
I will never make such commits. And I did.

This commit includes two little gems.

No. 1
A RedbeanPHP extension which allows to store revisions of a bean (structure and data) in a shadow bean.

No. 2
A not yet there kind of animation that highlights span elements within a .readme class in a reading or karaoke style fashion.